Complete + sim-verify the AXI kernel datapath (gemm_kernel)
Turn the Vitis RTL-kernel wrapper from a skeleton into a full, simulated
datapath: AXI4-Lite control (ap_ctrl_hs register map) + AXI4-master DMA that
reads A/B from global memory, feeds the systolic array with the systolic skew,
collects C, and writes it back. Verified in cocotb against golden with an AXI
memory-model slave (tb/test_kernel.py; `make kernel`), across 4x4, 2x6, 6x3,
8x8 and varying K.

Fixes a real ap_ctrl_hs bug caught in sim: ap_done was a 1-cycle pulse that the
host's polling could miss -> made it sticky (set on done, cleared on status
read). Exactly the class of bug that would otherwise cost hours on the ILA.

host/run_hw.py updated to the 32-bit-per-element memory layout.

+++ host/run_hw.py
@@ -39,11 +39,13 @@ def main():
     krnl = pyxrt.kernel(dev, uuid, "gemm_kernel")
 
     # --- данные (INT8), C — INT32 ---
+    # Раскладка памяти = как в rtl/gemm_kernel.sv: 32-битное слово НА ЭЛЕМЕНТ
+    # (INT8 со знаковым расширением в INT32). Плотная INT8-упаковка = оптимизация.
     rng = np.random.default_rng(0xBEEF)
     A = rng.integers(-128, 128, size=(args.M, args.K), dtype=np.int8)
     B = rng.integers(-128, 128, size=(args.K, args.N), dtype=np.int8)
-    a_bytes = A.tobytes()
-    b_bytes = B.tobytes()
+    a_bytes = A.astype(np.int32).tobytes()   # 4 байта/элемент
+    b_bytes = B.astype(np.int32).tobytes()
     c_nbytes = args.M * args.N * 4  # INT32
 
     # --- буферы в глобальной памяти (group_id по порядку аргументов kernel) ---
